==> 4.EFA_base_model.Rmd <==
---
title: "EFA crimes-base"
output: html_document
date: "2024-07-16"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
```{r}
library(tidyverse)
library(SMOTEWB)
library(smotefamily)
library(randomForest)
library(DBI)
library(glmnet)
library(caret)
library(purrr)
library(vcd)
library(DescTools)
library(knitr)
library(kableExtra)
library(stargazer)
library(psych)
library(FactoMineR)
library(factoextra)
```


```{r loading data}
data <- read.csv("/Users/valentinalaugeri/Documents/LSE/Diss/dissertation/data/imputed_cleaned_data.csv")
data <- data %>% 
	mutate(group = factor(group))
```

```{r}
data <- data %>%
  mutate(fdextort = factor(fdextort, levels = c(0, 1))) %>% 
	mutate(fdsmuggl = factor(fdsmuggl, levels = c(0, 1))) %>% 
	mutate(fdrob = factor(fdrob, levels = c(0, 1))) %>% 
	mutate(mul_bases = factor(mul_bases, levels = c(0, 1))) %>% 
	mutate(crim_degr_py = as.numeric(crim_degr_py)) %>% 
	mutate(endstr = factor(endstr, levels = c(0, 1, 2, 3, 4, 5))) %>% 
	mutate(end = factor(end, levels = c(0, 1))) %>% 
	mutate(ethn = factor(ethn, levels = c(0, 1))) %>% 
	mutate(EAP = factor(EAP, levels = c(0, 1))) %>% 
	mutate(ECA = factor(ECA, levels = c(0, 1))) %>% 
	mutate(LAC = factor(LAC, levels = c(0, 1))) %>% 
	mutate(MENA = factor(MENA, levels = c(0, 1))) %>% 
	mutate(North_America = factor(North_America, levels = c(0, 1))) %>% 
	mutate(SAS = factor(SAS, levels = c(0, 1))) %>% 
	mutate(SSA = factor(SSA, levels = c(0, 1))) %>% 
	mutate(nat = factor(nat, levels = c(0, 1))) %>% 
	mutate(right = factor(right, levels = c(0, 1))) %>% 
	mutate(ercsr = factor(ercsr, levels = c(0, 1))) %>% 
	mutate(pch = factor(pch, levels = c(0, 1))) %>% 
	mutate(sq = factor(sq, levels = c(0, 1))) %>% 
	mutate(tch = factor(tch, levels = c(0, 1))) %>% 
	mutate(public_service = factor(public_service, levels = c(0, 1))) %>% 
	mutate(fate_leader = factor(fate_leader, levels = c(0, 1, 2))) %>% 
	mutate(soc_service = factor(soc_service, levels = c(0, 1))) %>% 
	mutate(combined_size = factor(combined_size, levels = c("1", "2", "3", "4", "5"))) %>% 
  mutate(combined_reli = factor(combined_reli, levels = c(0, 1))) %>% 
  mutate(combined_left = factor(combined_left, levels = c(0, 1))) %>% 
  mutate(combined_hierarch = factor(combined_hierarch, levels = c(0, 1))) %>% 
  mutate(combined_drug_trafficking = factor(combined_drug_trafficking, levels = c(0, 1))) %>%
  mutate(combined_kidnap = factor(combined_kidnap, levels = c(0, 1))) %>%
  mutate(combined_state_sponsor = factor(combined_state_sponsor, levels = c(0, 1))) %>%
  mutate(combined_terrctrl = factor(combined_terrctrl, levels = c(0, 1)))
```

```{r}
data <- data %>% 
	mutate(combined_size = ifelse(combined_size == 1, 10, combined_size)) %>%
	mutate(combined_size = ifelse(combined_size == 2, 100, combined_size)) %>%
	mutate(combined_size = ifelse(combined_size == 3, 1000, combined_size)) %>%
	mutate(combined_size = ifelse(combined_size == 4, 5000, combined_size)) %>%
	mutate(combined_size = ifelse(combined_size == 1, 10000, combined_size)) %>%
	mutate(combined_size = as.numeric(combined_size))
```

```{r}
data <- data %>% 
	mutate(pure_religious = ifelse(combined_reli == 1 & (nat == 0 & ethn == 0 & combined_left == 0 & ercsr == 0 & right == 0), 1, 0)) %>% 
	mutate(pure_secular = ifelse((ethn == 1 & combined_reli == 0) |(combined_left == 1 & combined_reli == 0) |(ercsr == 1 & combined_reli ==0) |(right == 1 & combined_reli == 0), 1, 0)) %>% 
	mutate(secular_hybrid = ifelse(((ethn == 1 & combined_reli == 0) &(combined_left == 1 | ercsr == 1 | right == 1)) | 
														((combined_left == 1 & combined_reli == 0) &(ethn == 1 | ercsr == 1 | right == 1)) |
														((ercsr == 1 & combined_reli ==0) & (ethn == 1 |  combined_left == 1 | right == 1)) |
														((right == 1 & combined_reli == 0) & (ethn == 1 |  ercsr == 1 | combined_left == 1)), 1, 0)) %>% 
	mutate(hybrid_religious = ifelse(combined_reli == 1 & (ethn == 1 | combined_left == 1 | ercsr == 1 | right == 1), 1, 0))
```
```{r Scaling numeric variables + numeric factors}
numeric_vars <- data %>% select_if(is.numeric)
numeric_vars_standard <- scale(numeric_vars)
# replace in og data
data[, names(numeric_vars_standard)] <- numeric_vars_standard

for(i in names(data)){
	if(is.factor(i)){
		i <- as.numeric(i)
	}
}
data <- data %>% select(-X, -group, -combined_base)
data$combined_state_sponsor <- as.numeric(data$combined_state_sponsor)
data$combined_terrctrl <- as.numeric(data$combined_terrctrl)
```

```{r}
vars_model <- data %>% 
	mutate(combined_drug_trafficking = as.numeric(combined_drug_trafficking)) %>% 
	mutate(combined_kidnap = as.numeric(combined_kidnap)) %>% 
	mutate(fdsmuggl = as.numeric(fdsmuggl)) %>% 
	mutate(fdextort = as.numeric(fdextort)) %>% 
	mutate(fdrob = as.numeric(fdrob)) 
vars_model <- vars_model %>% 
	mutate("Drug Trafficking" = combined_drug_trafficking) %>% 
	mutate("Kindapping" = combined_kidnap) %>% 
	mutate("Smuggling" = fdsmuggl) %>% 
	mutate("Robbery" = fdrob) %>% 
	mutate("Extortion" = fdextort )
vars_model <- vars_model %>% 
	select(c("Drug Trafficking", "Kindapping", "Smuggling", "Robbery", "Extortion"))
```

```{r}
pca_result <- prcomp(vars_model, scale. = FALSE)

# Print the summary of the PCA result to see the explained variance
summary(pca_result)

# Visualize the results
# Scree plot
#png("PCA-base.png", width = 1600, height = 1200, res = 300)
fviz_eig(pca_result, addlabels = TRUE, ylim = c(0, 50), 
				 main = "Figure 2: Principal Component Analysis on Crimes")

#dev.off()
# Variable loadings plot
png("PCA-base-1.png", width = 2000, height = 1200, res = 300)
fviz_pca_var(pca_result,
             col.var = "contrib", # Color by contributions to the PC
             gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"),
             repel = TRUE,
						 title = "Figure 2: Principal Component Analysis on Crimes - Dimension Loadings", 
						 xlab = "Dimension 1 (74.9%)", 
						 ylab = "Dimension 2 (12.4%)", 
						 legend.title = "Contribution"
						 )        # Avoid text overlapping
dev.off()
# Individual plot
# fviz_pca_ind(pca_result,
#              col.ind = "cos2",    # Color by the quality of representation
#              gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"),
#              repel = TRUE)        # 
```


```{r}
# Perform factor analysis with 5 factors
fa_result <- fa(vars_model, nfactors = 2, rotate = "varimax")
fa.diagram(fa_result)

# Set up the PNG device to save the plot with higher resolution
png("EFA-base.png", width = 2000, height = 1200, res = 300)

# Create the factor analysis diagram with customization
fa.diagram(fa_result,
           main = "Figure 1: Exploratory Factor Analysis on Crimes", 
           rsize = 0.6, 
           edge.col = c("black", "red"),
           labels = rename_vars,
           cut = 0.4)  # cut-off for loading values displayed

# Close the PNG device to save the file
dev.off()
```
